chore: moved footer
